% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Raw Data set}
\format{
List
#' \describe{
\item{Country}{Pakistan.}
\item{admin1_District}{Data frame in wide format.}
\item{metadata}{Description of the variables.}
\item{context}{context of the variables - in case of the need to reverse color scheme}
}
}
\source{
Author prepared based on data inputs by CCDR TEAM - World Bank
}
\usage{
data
}
\description{
Raw Data set
}
\keyword{datasets}
