% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pca}
\alias{data_pca}
\title{Final Data set to be used in PCA algorithm}
\format{
Wide data frame - tibble
}
\source{
CCDR TEAM - World Bank
}
\usage{
data_pca
}
\description{
Final Data set to be used in PCA algorithm
}
\keyword{datasets}
