% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts}
\alias{districts}
\title{District Names - Pakistan}
\format{
tibble
}
\source{
Derived from Shape file
}
\usage{
districts
}
\description{
District Names - Pakistan
}
\keyword{datasets}
