% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pak_district}
\alias{pak_district}
\title{District Shapefile - Pakistan}
\format{
Data frame with columns - prepared and oriented
\describe{
\item{admin0Pcod}{Country Name.}
\item{admin1Pcod}{District Code.}
\item{admin1Name}{District Name.}
\item{geometry}{Geometry Coords.}
}
}
\source{
<World Bank>
}
\usage{
pak_district
}
\description{
District Shapefile - Pakistan
}
\keyword{datasets}
